Created components for bracket sections, podium fixed, data structure cleaned up
