+++ components/bracket-matches.tsx
@@ -1,0 +1,140 @@
+import Image from "next/image"
+import { TwitchEmbed } from "react-twitch-embed"
+import { Matches } from "types/playoffs"
+
+import { addRank, eloColor, timeFormat } from "@/lib/utils"
+import { Card, CardContent, CardTitle } from "@/components/ui/card"
+
+interface BracketMatchesProps {
+  matches: Matches[]
+  nextNonActiveMatch: string
+}
+
+export default function BracketMatches(props: BracketMatchesProps) {
+  return (
+    <>
+      {props.matches.map((match) =>
+        match.state === "ACTIVE" ? (
+          <>
+            <Card className="w-1/2 rounded-lg border-2 border-green-400">
+              <CardTitle className="mt-10 text-center">
+                Round 1 - Match 1
+              </CardTitle>
+              <CardContent>
+                <div className="flex justify-between space-x-4 p-4 text-center">
+                  <div>
+                    <Image
+                      src={`https://mc-heads.net/body/${match.participants[0].name}`}
+                      alt="Players skin"
+                      width={120}
+                      height={720}
+                    />
+                    <div className="mt-3">
+                      <h1>{match.participants[0].name}</h1>
+                      <h1
+                        className={`${eloColor(
+                          match.participants[0]?.season_elo
+                        )}`}
+                      >
+                        {match.participants[0]?.season_elo}
+                      </h1>
+                      <h1>{timeFormat(match.participants[0].personal_best)}</h1>
+                    </div>
+                  </div>
+                  <h1 className="my-auto flex justify-center text-4xl font-semibold">
+                    VS
+                  </h1>
+                  <div>
+                    <Image
+                      src={`https://mc-heads.net/body/${match.participants[1].name}/left`}
+                      alt="Players skin"
+                      width={120}
+                      height={720}
+                    />
+                    <div className="mt-3">
+                      <h1>{match.participants[1].name}</h1>
+                      <h1
+                        className={`${eloColor(
+                          match.participants[1]?.season_elo
+                        )}`}
+                      >
+                        {match.participants[1]?.season_elo}
+                      </h1>
+                      <h1>{timeFormat(match.participants[1].personal_best)}</h1>
+                    </div>
+                  </div>
+                </div>
+              </CardContent>
+            </Card>
+            <TwitchEmbed
+              channel="Feinberg"
+              autoplay
+              withChat={false}
+              darkMode
+              hideControls
+              className="mt-2"
+            />
+          </>
+        ) : match.state !== "ACTIVE" &&
+          match.state !== "DONE" &&
+          match.startTime === props.nextNonActiveMatch ? (
+          <>
+            {" "}
+            <Card className="w-full rounded-lg border-2 border-green-400">
+              <CardTitle className="mt-10 text-center">
+                {`UP NEXT @ ${new Date(
+                  parseInt(match.startTime) * 1000
+                ).toLocaleTimeString([], {
+                  timeZoneName: "short",
+                  hour: "2-digit",
+                  minute: "2-digit",
+                })}`}
+                <br />
+                {`Round - ${match.tournamentRoundText} `}
+              </CardTitle>
+              <CardContent>
+                <div className="flex justify-between space-x-4 p-4 text-center">
+                  <div>
+                    <Image
+                      src={`https://mc-heads.net/body/${match.participants[0].name}`}
+                      alt="Players skin"
+                      width={120}
+                      height={720}
+                    />
+                    <div className="mt-3">
+                      <h1>{match.participants[0].name}</h1>
+                      <h1 className={`${eloColor(2000)}`}>
+                        {`2000 - ${addRank(2000)}`}
+                      </h1>
+                      <h1>{timeFormat(match.participants[0].personal_best)}</h1>
+                    </div>
+                  </div>
+                  <h1 className="my-auto flex justify-center text-4xl font-semibold">
+                    VS
+                  </h1>
+                  <div>
+                    <Image
+                      src={`https://mc-heads.net/body/${match.participants[1].name}/left`}
+                      alt="Players skin"
+                      width={120}
+                      height={720}
+                    />
+                    <div className="mt-3">
+                      <h1>{match.participants[1].name}</h1>
+                      <h1 className={`${eloColor(2000)}`}>
+                        {`2000 - ${addRank(2000)}`}
+                      </h1>
+                      <h1>{timeFormat(match.participants[1].personal_best)}</h1>
+                    </div>
+                  </div>
+                </div>
+              </CardContent>
+            </Card>
+          </>
+        ) : (
+          " "
+        )
+      )}
+    </>
+  )
+}
