+++ package.json
@@ -19,7 +19,7 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@lowgy/react-tournament-brackets": "^1.0.6",
+    "@lowgy/react-tournament-brackets": "^1.0.11",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-scroll-area": "^1.0.4",
@@ -32,7 +32,6 @@
     "@tanstack/react-table": "^8.9.1",
     "class-variance-authority": "^0.4.0",
     "clsx": "^1.2.1",
-    "framer-motion": "^10.16.4",
     "lucide-react": "^0.105.0-alpha.4",
     "next": "^13.4.4",
     "next-themes": "^0.2.1",
